refactor(core): extract maintenance/relink.ts — raw-event relink plan/apply family
